requirements.txt
ag2==0.9.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
blinker==1.9.0
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.5
curl_cffi==0.11.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
diskcache==5.6.3
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.0
fastmcp==2.10.2
filetype==1.2.0
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.5.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.101.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.24.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-google-vertexai==2.0.27
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.72
langsmith==0.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.10.1
mdurl==0.1.2
multidict==6.6.3
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
numexpr==2.11.0
numpy==2.3.1
openapi-pydantic==0.5.1
openinference-instrumentation==0.1.34
openinference-instrumentation-autogen==0.1.9
openinference-instrumentation-langchain==0.1.46
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-instrumentation==0.55b1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.1
tenacity==8.5.0
termcolor==3.1.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.19
uvicorn==0.35.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.23.0

fastapi
uvicorn
pypdf
python-docx
chromadb
langchain
langchain-google-genai
python-dotenv
camelot-py
openpyxl
